requirements.txt
Django==2.2
django-celery-beat
django-celery-results
psycopg2
Celery==4.4.0
gunicorn
python-telegram-bot
cached-property